requirements.txt
nltk
noaa_sdk
speechrecognition
pyaiml
requests
